Built an index optimization
The optimization works with a list of cube states that we know are
solvable. While traversing the graph, if one of these states is found,
then we clear the queue and start from that point. It's a proof of
concept with about ~100,000 states in the index but it extended the
number of face turns to shuffle a cube from 6 to 16. At 16, not every
cube is solvable, but we know within 2 seconds if it's solvable or not.